refactor: separate project modules
