+++ src/ape/managers/project/__init__.py
@@ -1,0 +1,7 @@
+from ape.managers.project.dependency import DependencyManager
+from ape.managers.project.manager import ProjectManager
+
+__all__ = [
+    "DependencyManager",
+    "ProjectManager",
+]
