+++ src/ape/managers/project/manager.py
@@ -1,242 +1,14 @@
-import json
 import shutil
-import tempfile
 from pathlib import Path
 from typing import Dict, List, Optional, Type, Union
 
-import yaml
-from ethpm_types import Compiler, ContractType, PackageManifest
-from ethpm_types.utils import compute_checksum
+from ethpm_types import Compiler, ContractType
 
-from ape.api.projects import DependencyAPI, ProjectAPI
+from ape.api import DependencyAPI, ProjectAPI
 from ape.contracts import ContractContainer
 from ape.exceptions import ProjectError
-from ape.logging import logger
-from ape.utils import (
-    ManagerAccessMixin,
-    cached_property,
-    get_all_files_in_directory,
-    get_relative_path,
-    github_client,
-)
-
-from .base import BaseManager
-from .config import CONFIG_FILE_NAME
-
-
-class ApeProject(ProjectAPI):
-    """
-    The default implementation of the :class:`~ape.api.projects.ProjectAPI`.
-    By default, the `:class:`~ape.managers.project.ProjectManager` uses an
-    ``ApeProject`` at the current-working directory.
-    """
-
-    @property
-    def is_valid(self) -> bool:
-        if (self.path / CONFIG_FILE_NAME).exists():
-            return True
-
-        logger.warning(
-            f"'{self.path.name}' is not an 'ApeProject', but attempting to process as one."
-        )
-
-        # NOTE: We always return True as a last-chance attempt because it often
-        # works anyway and prevents unnecessary plugin requirements.
-        return True
-
-    @property
-    def sources(self) -> List[Path]:
-        """
-        All the source files in the project.
-        Excludes files with extensions that don't have a registered compiler.
-
-        Returns:
-            List[pathlib.Path]: A list of a source file paths in the project.
-        """
-        files: List[Path] = []
-
-        if not self.contracts_folder.exists():
-            return files
-
-        for extension in self.compiler_manager.registered_compilers:
-            files.extend(self.contracts_folder.rglob(f"*{extension}"))
-
-        return files
-
-    def create_manifest(
-        self, file_paths: Optional[List[Path]] = None, use_cache: bool = True
-    ) -> PackageManifest:
-        # Create a config file if one doesn't exist to forward values from
-        # the root project's 'ape-config.yaml' 'dependencies:' config.
-        config_file = self.path / CONFIG_FILE_NAME
-        if not config_file.exists():
-            config_data = {}
-            if self.name:
-                config_data["name"] = self.name
-            if self.version:
-                config_data["version"] = self.version
-            if self.contracts_folder.name != "contracts":
-                # Only sets when not default.
-                config_data["contracts_folder"] = self.contracts_folder.name
-            if config_data:
-                with open(config_file, "w") as f:
-                    yaml.safe_dump(config_data, f)
-
-        # Load a cached or clean manifest (to use for caching)
-        if self.cached_manifest and use_cache:
-            manifest = self.cached_manifest
-        else:
-            manifest = PackageManifest()
-
-            if self.manifest_cachefile.exists():
-                self.manifest_cachefile.unlink()
-
-        cached_sources = manifest.sources or {}
-        cached_contract_types = manifest.contract_types or {}
-        sources = {s for s in self.sources if s in file_paths} if file_paths else set(self.sources)
-
-        # Filter out deleted sources
-        deleted_source_ids = cached_sources.keys() - set(
-            map(str, [get_relative_path(s, self.contracts_folder) for s in sources])
-        )
-        contract_types = {
-            name: contract_type
-            for name, contract_type in cached_contract_types.items()
-            if contract_type.source_id not in deleted_source_ids
-        }
-
-        def file_needs_compiling(source: Path) -> bool:
-            path = str(get_relative_path(source, self.contracts_folder))
-
-            if path not in cached_sources:
-                return True  # New file added
-
-            cached = cached_sources[path]
-            cached.compute_checksum(algorithm="md5")
-            assert cached.checksum  # For mypy
-
-            source_file = self.contracts_folder / source
-            checksum = compute_checksum(
-                source_file.read_bytes(),
-                algorithm=cached.checksum.algorithm,
-            )
-
-            return checksum != cached.checksum.hash  # Contents changed
-
-        # NOTE: Filter by checksum to only update what's needed
-        needs_compiling = list(filter(file_needs_compiling, sources))
-
-        # Set the context in case compiling a dependency (or anything outside the root project).
-        with self.config_manager.using_project(self.path, contracts_folder=self.contracts_folder):
-            compiled_contract_types = self.compiler_manager.compile(needs_compiling)
-            contract_types.update(compiled_contract_types)
-
-            # NOTE: Update contract types & re-calculate source code entries in manifest
-            sources = (
-                {s for s in self.sources if s in file_paths} if file_paths else set(self.sources)
-            )
-
-            dependencies = {c for c in get_all_files_in_directory(self.contracts_folder / ".cache")}
-            for contract in dependencies:
-                sources.add(contract)
-
-            manifest = self._create_manifest(
-                sources,
-                self.contracts_folder,
-                contract_types,
-                initial_manifest=manifest,
-            )
-
-            # NOTE: Cache the updated manifest to disk (so ``self.cached_manifest`` reads next time)
-            self.manifest_cachefile.write_text(json.dumps(manifest.dict()))
-            return manifest
-
-
-class LocalDependency(DependencyAPI):
-    """
-    A dependency that is already downloaded on the local machine.
-
-    Config example::
-
-        dependencies:
-          - name: Dependency
-            local: path/to/dependency
-    """
-
-    local: str
-
-    @property
-    def path(self) -> Path:
-        given_path = Path(self.local)
-        if not given_path.exists():
-            raise ProjectError(f"No project exists at path '{given_path}'.")
-
-        return given_path
-
-    @property
-    def version_id(self) -> str:
-        return "local"
-
-    def extract_manifest(self) -> PackageManifest:
-        return self._extract_local_manifest(self.path)
-
-
-class GithubDependency(DependencyAPI):
-    """
-    A dependency from Github. Use the ``github`` key in your ``dependencies:``
-    section of your ``ape-config.yaml`` file to declare a dependency from GitHub.
-
-    Config example::
-
-        dependencies:
-          - name: OpenZeppelin
-            github: OpenZeppelin/openzeppelin-contracts
-            version: 4.4.0
-    """
-
-    github: str
-    """
-    The Github repo ID e.g. the organization name followed by the repo name,
-    such as ``dapphub/erc20``.
-    """
-
-    branch: Optional[str] = None
-    """
-    The branch to use. **NOTE**: Will be ignored if given a version.
-    """
-
-    @property
-    def version_id(self) -> str:
-        if self.branch:
-            return self.branch
-
-        if self.version and self.version != "latest":
-            return self.version
-
-        latest_release = github_client.get_release(self.github, "latest")
-        return latest_release.tag_name
-
-    def __repr__(self):
-        return f"<{self.__class__.__name__} github={self.github}>"
-
-    def extract_manifest(self) -> PackageManifest:
-        if self.cached_manifest:
-            # Already downloaded
-            return self.cached_manifest
-
-        with tempfile.TemporaryDirectory() as temp_dir:
-            temp_project_path = (Path(temp_dir) / self.name).resolve()
-            temp_project_path.mkdir(exist_ok=True, parents=True)
-
-            if self.branch:
-                github_client.clone_repo(self.github, temp_project_path, branch=self.branch)
-
-            else:
-                github_client.download_package(
-                    self.github, self.version or "latest", temp_project_path
-                )
-
-            return self._extract_local_manifest(temp_project_path)
+from ape.managers.base import BaseManager
+from ape.managers.project.types import ApeProject
 
 
 class ProjectManager(BaseManager):
@@ -653,32 +425,3 @@ class ProjectManager(BaseManager):
     #     TODO: Clean up manifest and minify it
     #     TODO: Publish sources to IPFS and replace with CIDs
     #     TODO: Publish to IPFS
-
-
-class DependencyManager(ManagerAccessMixin):
-    DATA_FOLDER: Path
-
-    def __init__(self, data_folder: Path):
-        self.DATA_FOLDER = data_folder
-
-    @cached_property
-    def dependency_types(self) -> Dict[str, Type[DependencyAPI]]:
-        dependency_classes = {
-            "github": GithubDependency,
-            "local": LocalDependency,
-        }
-
-        for _, (config_key, dependency_class) in self.plugin_manager.dependencies:
-            dependency_classes[config_key] = dependency_class
-
-        return dependency_classes  # type: ignore
-
-    def decode_dependency(self, config_dependency_data: Dict) -> DependencyAPI:
-        for key, dependency_cls in self.dependency_types.items():
-            if key in config_dependency_data:
-                return dependency_cls(
-                    **config_dependency_data,
-                )  # type: ignore
-
-        dep_id = config_dependency_data.get("name", json.dumps(config_dependency_data))
-        raise ProjectError(f"No installed dependency API that supports '{dep_id}'.")
